+++ api/src/app/agent_tasks/profile_analyzer_agent.py
@@ -1,8 +1,14 @@
-from agents import Agent, output_guardrail, GuardrailFunctionOutput, WebSearchTool
+"""
+Module: agent_tasks.profile_analyzer_agent
+
+Defines the Profile Analyzer agent and its data models (input/output schemas).
+"""
+from agents import Agent
 from agents.model_settings import ModelSettings
-from typing import List, Optional
-from typing_extensions import TypedDict
+from agents.tool import WebSearchTool
 from pydantic import BaseModel
+from typing import List
+from typing_extensions import TypedDict
 
 class ProfileModel(BaseModel):
     display_name: str
@@ -47,6 +53,7 @@ class ProfileAnalyzerOut(TypedDict):
     output_type: str
     report: ProfileAnalyzerReport
 
+# Definition of the LLM agent
 profile_analyzer_agent = Agent(
     name="Profile Analyzer",
     model="gpt-4.1-mini",
