--- codex/tasks/0_api_refactor_t5.md
@@ -1,3 +1,0 @@
-## 
-
-## Task 5
