--- codex/tasks/0_api_refactor_t6.md
@@ -1,3 +1,0 @@
-## 
-
-## Task 6
